% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_file_github.R
\name{show_file_github}
\alias{show_file_github}
\title{Show a file from github}
\usage{
show_file_github(repo, file, ...)
}
\arguments{
\item{repo}{a github repository}

\item{file}{a file within the github repository}

\item{...}{argumetns passed down to \code{\link[=show_file]{show_file()}}}
}
\description{
Show a file from github
}
\examples{
\dontrun{
show_file_github("GregorDeCillia/showdown", "DESCRIPTION")
}

}
