% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_file.R
\name{show_file}
\alias{show_file}
\title{show contents of a file with syntax highlighting}
\usage{
show_file(file, language = file_ext(file))
}
\arguments{
\item{file}{a path to a file to show}

\item{language}{the language used for syntax higlighing}
}
\description{
show contents of a file with syntax highlighting
}
